+++ CHANGELOG.md
@@ -1,6 +1,17 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## 1.1.2 - 2017-11-07
+
+### Refactored
+- now using `OOP` approach
+
+### Fixed
+- `output` files names concatenation with `+`
+
+### Removed
+- additional labels algorithm (to be added)
+
 ## 1.1.1 - 2017-11-04
 
 ### Added